+++ frontend/javascripts/oxalis/default_state.js
@@ -76,6 +76,7 @@ const defaultState: OxalisState = {
   },
   temporaryConfiguration: {
     viewMode: Constants.MODE_PLANE_TRACING,
+    histogramData: {},
     flightmodeRecording: false,
     controlMode: ControlModeEnum.VIEW,
     mousePosition: null,