+++ frontend/javascripts/oxalis/model/reducers/settings_reducer.js
@@ -11,6 +11,7 @@ import {
   type StateShape2,
 } from "oxalis/model/helpers/deep_update";
 import { clamp } from "libs/utils";
+import { getDefaultIntensityRangeOfLayer } from "oxalis/model/accessors/dataset_accessor";
 import { userSettings } from "libs/user_settings.schema";
 
 //
@@ -86,6 +87,7 @@ function SettingsReducer(state: OxalisState, action: Action): OxalisState {
       const layerSettingsDefaults = _.transform(
         state.dataset.dataSource.dataLayers,
         (result, layer) => {
+          const intensityRange = getDefaultIntensityRangeOfLayer(state.dataset, layer.name);
           result[layer.name] = Object.assign(
             {},
             {
@@ -93,7 +95,7 @@ function SettingsReducer(state: OxalisState, action: Action): OxalisState {
               contrast: 1,
               color: [255, 255, 255],
               alpha: layer.category === "color" ? 100 : 20,
-              intensityRange: [0, 255],
+              intensityRange,
               isDisabled: false,
               isInverted: false,
             },
@@ -126,6 +128,9 @@ function SettingsReducer(state: OxalisState, action: Action): OxalisState {
       } else {
         return state;
       }
+    }
+    case "SET_HISTOGRAM_DATA": {
+      return updateTemporaryConfig(state, { histogramData: action.histogramData });
     }
     case "SET_FLIGHTMODE_RECORDING": {
       return updateTemporaryConfig(state, { flightmodeRecording: action.value });