+++ frontend/javascripts/oxalis/view/settings/dataset_settings_view.js
@@ -11,7 +11,7 @@ import _ from "lodash";
 import { V3 } from "libs/mjs";
 import api from "oxalis/api/internal_api";
 
-import type { APIDataset, APIHistogramData } from "admin/api_flow_types";
+import type { APIDataset } from "admin/api_flow_types";
 import { AsyncIconButton } from "components/async_clickables";
 import {
   SwitchSetting,
@@ -19,13 +19,14 @@ import {
   DropdownSetting,
   ColorSetting,
 } from "oxalis/view/settings/setting_input_views";
-import { findDataPositionForLayer, getHistogramForLayer, clearCache } from "admin/admin_rest_api";
+import { findDataPositionForLayer, clearCache } from "admin/admin_rest_api";
 import { getGpuFactorsWithLabels } from "oxalis/model/bucket_data_handling/data_rendering_logic";
 import { getMaxZoomValueForResolution } from "oxalis/model/accessors/flycam_accessor";
 import {
   hasSegmentation,
   getElementClass,
   getLayerBoundaries,
+  getDefaultIntensityRangeOfLayer,
 } from "oxalis/model/accessors/dataset_accessor";
 import { setPositionAction, setZoomStepAction } from "oxalis/model/actions/flycam_actions";
 import {
@@ -39,6 +40,7 @@ import Store, {
   type DatasetLayerConfiguration,
   type OxalisState,
   type UserConfiguration,
+  type HistogramDataForAllLayers,
   type Tracing,
 } from "oxalis/store";
 import Toast from "libs/toast";
@@ -67,6 +69,7 @@ type DatasetSettingsProps = {|
     value: any,
   ) => void,
   viewMode: ViewMode,
+  histogramData: HistogramDataForAllLayers,
   controlMode: ControlMode,
   hasSegmentation: boolean,
   onSetPosition: Vector3 => void,
@@ -75,36 +78,7 @@ type DatasetSettingsProps = {|
   tracing: Tracing,
 |};
 
-type State = {
-  histograms: { string: APIHistogramData },
-};
-
-class DatasetSettings extends React.PureComponent<DatasetSettingsProps, State> {
-  state = {
-    histograms: {},
-  };
-
-  componentDidMount() {
-    this.loadAllHistograms();
-  }
-
-  loadAllHistograms = async () => {
-    const { layers } = this.props.datasetConfiguration;
-    const histograms: { string: APIHistogramData } = {};
-    const histogramPromises = Object.keys(layers).map(async layerName => {
-      const data = await getHistogramForLayer(
-        this.props.dataset.dataStore.url,
-        this.props.dataset,
-        layerName,
-      );
-      histograms[layerName] = data;
-      return data;
-    });
-    // Waiting for all Promises to be resolved.
-    await Promise.all(histogramPromises);
-    this.setState({ histograms });
-  };
-
+class DatasetSettings extends React.PureComponent<DatasetSettingsProps> {
   getFindDataButton = (layerName: string, isDisabled: boolean, isColorLayer: boolean) => {
     let tooltipText = isDisabled
       ? "You cannot search for data when the layer is disabled."
@@ -119,7 +93,7 @@ class DatasetSettings extends React.PureComponent<DatasetSettingsProps, State> {
       } else if (volume && !volume.fallbackLayer && !isDisabled) {
         isDisabled = true;
         tooltipText =
-          "You do not have a fallback layer for this segmentation layer. It is only possible to search in fallback layers";
+          "You do not have a fallback layer for this segmentation layer. It is only possible to search in fallback layers.";
       }
     }
 
@@ -202,17 +176,26 @@ class DatasetSettings extends React.PureComponent<DatasetSettingsProps, State> {
 
   getHistogram = (layerName: string, layer: DatasetLayerConfiguration) => {
     const { intensityRange } = layer;
-    let histograms = [
-      { numberOfElements: 256, elementCounts: new Array(256).fill(0), min: 0, max: 255 },
-    ];
-    if (this.state.histograms && this.state.histograms[layerName]) {
-      histograms = this.state.histograms[layerName];
+    const defaultIntensityRange = getDefaultIntensityRangeOfLayer(this.props.dataset, layerName);
+    const highestRangeValue = defaultIntensityRange[1];
+    let histograms = [];
+    if (this.props.histogramData && this.props.histogramData[layerName]) {
+      histograms = this.props.histogramData[layerName];
+    } else {
+      histograms = [
+        {
+          numberOfElements: 0,
+          elementCounts: [],
+          min: 0,
+          max: highestRangeValue,
+        },
+      ];
     }
     return (
       <Histogram
         data={histograms}
-        min={intensityRange[0]}
-        max={intensityRange[1]}
+        intensityRangeMin={intensityRange[0]}
+        intensityRangeMax={intensityRange[1]}
         layerName={layerName}
       />
     );
@@ -523,6 +506,7 @@ const mapStateToProps = (state: OxalisState) => ({
   userConfiguration: state.userConfiguration,
   datasetConfiguration: state.datasetConfiguration,
   viewMode: state.temporaryConfiguration.viewMode,
+  histogramData: state.temporaryConfiguration.histogramData,
   controlMode: state.temporaryConfiguration.controlMode,
   dataset: state.dataset,
   hasSegmentation: hasSegmentation(state.dataset),